+++ NEWS.rst
@@ -6,6 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
+* Major: Hardcoded ``packet_size`` and ``sync_byte`` as constexpr static
+  varibles.
 * Major: Removed ``set_verify`` from packetizer.
 * Major: Removed ``set_on_data`` from packetizer - this is not set via the
   contructor.